backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
ta==0.10.2
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0

